Add files via upload
This file contains text classifier code using NLTK and TensorFlow . The appitude questions can be classified into their category. For example we have categories like probability,permutation & combination and etc. So the code classifies the question to its respective category.
The training data is in the Json format and can be increased as per the application. Along with the python code I have also uploaded the json file for reference.